Preliminary Replanning Ability
Very major refactoring of the GraspStateMachine logic
to enable online trajectory replanning. Now there is a new
node called polynomial_generator_node that generates a
trajectory based on the current state. The resulting
polynomial is fed to the state machine node for following. This
required many related changes. Currently the replanning works
although the logic for sending intermediate waypoints from
the state machine to the planner is not implemented yet.